fix : Monster 로직 수정, MonsterFactory 분리
- Monster가 선택 됐는지를 판단하는 bool값을 추가했습니다.
- MonsterFactory를 BattleManager에서 분리해서 작성해뒀습니다.
